net: lib: nrf_cloud: fix build issue with FOTA poll support
The nrf_cloud_rest_fota sample failed to build when JITP
was enabled.
Update the nrf_cloud_fota_poll module to allow for the
pending job type to be accessed by the application.
Moved fota_download_init into nrf_cloud_fota_poll_init.
Added function comments and moved into nrf_cloud_fota_poll.h

Increased AT host stack size for the nrf_cloud_rest_fota
sample.

IRIS-8444
