+++ doc/nrf/releases_and_maturity/releases/release-notes-changelog.rst
@@ -589,6 +589,7 @@ Cellular samples
 
   * Added credential check before connecting to network.
   * Changed the sample use the functions in the :file:`nrf_cloud_fota_poll.c` and :file:`nrf_cloud_fota_common.c` files.
+  * Increased the :kconfig:option:`CONFIG_AT_HOST_STACK_SIZE` Kconfig option to 2048 bytes since nRF Cloud credentials are sometimes longer than 1024 bytes.
 
 * :ref:`nrf_cloud_rest_cell_pos_sample` sample:
 
@@ -1180,7 +1181,10 @@ Libraries for networking
 
 * :ref:`lib_nrf_cloud_fota` library:
 
-  * Added the :file:`nrf_cloud_fota_poll.c` file to consolidate the FOTA polling code from the :ref:`nrf_cloud_multi_service` and :ref:`nrf_cloud_rest_fota` samples.
+  * Added:
+
+    * The :file:`nrf_cloud_fota_poll.c` file to consolidate the FOTA polling code from the :ref:`nrf_cloud_multi_service` and :ref:`nrf_cloud_rest_fota` samples.
+    * The :file:`nrf_cloud_fota_poll.h` file.
 
 * :ref:`lib_mqtt_helper` library:
