+++ samples/cellular/nrf_cloud_multi_service/src/fota_support_coap.h
@@ -8,8 +8,6 @@
 #define FOTA_SUPPORT_COAP_H
 
 int coap_fota_init(void);
-int coap_fota_begin(void);
-
 int coap_fota_thread_fn(void);
 
 #endif /* FOTA_SUPPORT_COAP_H */
