+++ samples/cellular/nrf_cloud_rest_fota/prj.conf
@@ -48,8 +48,10 @@ CONFIG_NRF_MODEM_LIB=y
 CONFIG_LTE_LINK_CONTROL=y
 
 # AT Host library - Used to send AT commands directy from an UART terminal and to allow
-#		    integration with nRF Connect for Desktop LTE Link monitor application.
+#		    integration with nRF Connect for Desktop applications.
 CONFIG_AT_HOST_LIBRARY=y
+# Extended AT host stack size since some nrf_cloud credentials are longer than 1024 bytes.
+CONFIG_AT_HOST_STACK_SIZE=2048
 
 # Modem info
 CONFIG_MODEM_INFO=y
