+++ subsys/net/lib/nrf_cloud/src/nrf_cloud_fota_poll.c
@@ -6,6 +6,7 @@
 
 #include <zephyr/kernel.h>
 #include <net/nrf_cloud.h>
+#include <net/nrf_cloud_fota_poll.h>
 #if defined(CONFIG_NRF_CLOUD_COAP)
 #include <net/nrf_cloud_coap.h>
 #endif
@@ -40,7 +41,7 @@ static K_SEM_DEFINE(fota_download_sem, 0, 1);
 static struct nrf_cloud_fota_job_info job;
 
 /* Pending job information saved to flash */
-static struct nrf_cloud_settings_fota_job pending_job;
+static struct nrf_cloud_settings_fota_job pending_job = { .type = NRF_CLOUD_FOTA_TYPE__INVALID };
 
 /* FOTA job status */
 static enum nrf_cloud_fota_status fota_status = NRF_CLOUD_FOTA_QUEUED;
@@ -215,14 +216,18 @@ int nrf_cloud_fota_poll_init(struct nrf_cloud_fota_poll_ctx *ctx)
 	ctx->full_modem_fota_supported = true;
 #endif
 
+	err = fota_download_init(http_fota_dl_handler);
+	if (err) {
+		LOG_ERR("Failed to initialize FOTA download, error: %d", err);
+		return err;
+	}
+
 	initialized = true;
 	return 0;
 }
 
-int nrf_cloud_fota_poll_start(struct nrf_cloud_fota_poll_ctx *ctx)
+int nrf_cloud_fota_poll_process_pending(struct nrf_cloud_fota_poll_ctx *ctx)
 {
-	int err;
-
 	if (!check_ctx(ctx) || !initialized) {
 		return -EINVAL;
 	}
@@ -230,13 +235,7 @@ int nrf_cloud_fota_poll_start(struct nrf_cloud_fota_poll_ctx *ctx)
 	/* This function may perform a reboot if a FOTA update is in progress */
 	process_pending_job(ctx);
 
-	err = fota_download_init(http_fota_dl_handler);
-	if (err) {
-		LOG_ERR("Failed to initialize FOTA download, error: %d", err);
-		return err;
-	}
-
-	return 0;
+	return pending_job.type;
 }
 
 static bool validate_in_progress_job(void)
@@ -487,6 +486,8 @@ int nrf_cloud_fota_poll_process(struct nrf_cloud_fota_poll_ctx *ctx)
 	 */
 	if (fota_status == NRF_CLOUD_FOTA_SUCCEEDED) {
 		handle_download_succeeded_and_reboot(ctx);
+		/* Application was expected to reboot... */
+		return -EBUSY;
 	}
 
 	/* Job was not successful, send status to nRF Cloud */
@@ -496,6 +497,5 @@ int nrf_cloud_fota_poll_process(struct nrf_cloud_fota_poll_ctx *ctx)
 	}
 
 	wait_after_job_update();
-
-	return err;
+	return -EFAULT;
 }
